operator: Fix operator hive population with default config
LBIPAM nil-deref'd when k8s configuration was not provided, fix
by adding a nil check. While at it, move lbipam cell definitions
to cell.go.

Add test to operator to check that the operator hive can be populated.
